+++ Test/SmallCheck/Series/Instances.hs
@@ -5,24 +5,30 @@
 {-|
 'Serial' instances are provided for the following types:
 
+* 'Data.Word'
+* 'Data.Word8'
 * 'Data.ByteString.ByteString'
 * 'Data.ByteString.Lazy.ByteString'
 * 'Data.Text.Text'
 * 'Data.Text.Lazy.Text'
+* 'Data.Text.Lazy.Text'
+* 'Data.Map.Map'
 
-By default the most economical but less exhaustive series are provided.
-You may want to create your own instances for your own tests if this setting
-doesn't apply for your own tests. Check the utilities in the other modules
-in this package to create custom 'Series'.
+By default the most exhaustive series are provided which can lead to
+combinatorial explosion if you are not careful. In such case, you may want to
+use the functions provided in the other modules in this package to create your
+own custom series.
 
-Make sure the module where you import these instances is not meant to be
-imported, otherwise the orphan instances might create issues.
+Make sure the module where you import these instances will not be imported,
+otherwise you might get conflicts between orphan instances defined in different
+modules.
 -}
 module Test.SmallCheck.Series.Instances () where
 
 #if !MIN_VERSION_base(4,8,0)
 import Control.Applicative ((<$>))
 #endif
+import Data.Word (Word8)
 import qualified Data.ByteString as B
 import qualified Data.ByteString.Lazy as BL
 import qualified Data.Text as T
@@ -31,14 +37,14 @@ import Data.Map
 import qualified Data.Map as Map
 import Test.SmallCheck.Series
 
+instance Monad m => Serial m Word where series = positive
+instance Monad m => CoSerial m Word where coseries = copositive
 
-import Test.SmallCheck.Series.ByteString as Series.ByteString
-import Test.SmallCheck.Series.ByteString.Lazy as Series.ByteString.Lazy
-import Test.SmallCheck.Series.Text as Series.Text
-import Test.SmallCheck.Series.Text.Lazy as Series.Text.Lazy
+instance Monad m => Serial m Word8 where series = positive
+instance Monad m => CoSerial m Word8 where coseries = copositive
 
 instance Monad m => Serial m B.ByteString where
-    series = Series.ByteString.replicateA
+    series = cons0 B.empty \/ cons2 B.cons
 instance Monad m => CoSerial m B.ByteString where
     coseries rs =
         alts0 rs >>- \y ->
@@ -48,7 +54,7 @@ instance Monad m => CoSerial m B.ByteString where
                 Just (b,bs') -> f (B.singleton b) bs'
 
 instance Monad m => Serial m BL.ByteString where
-    series = Series.ByteString.Lazy.replicateA
+    series = cons0 BL.empty \/ cons2 BL.cons
 instance Monad m => CoSerial m BL.ByteString where
     coseries rs =
         alts0 rs >>- \y ->
@@ -58,7 +64,7 @@ instance Monad m => CoSerial m BL.ByteString where
                 Just (b,bs') -> f (BL.singleton b) bs'
 
 instance Monad m => Serial m T.Text where
-    series = Series.Text.replicateA
+    series = cons0 T.empty \/ cons2 T.cons
 instance Monad m => CoSerial m T.Text where
     coseries rs =
         alts0 rs >>- \y ->
@@ -68,7 +74,7 @@ instance Monad m => CoSerial m T.Text where
                 Just (b,bs') -> f (T.singleton b) bs'
 
 instance Monad m => Serial m TL.Text where
-    series = Series.Text.Lazy.replicateA
+    series = cons0 TL.empty \/ cons2 TL.cons
 instance Monad m => CoSerial m TL.Text where
     coseries rs =
         alts0 rs >>- \y ->
@@ -90,3 +96,16 @@ instance (Ord k, CoSerial m k, CoSerial m v) => CoSerial m (Map k v) where
         pop m = case Map.toList m of
                      [] -> Nothing
                      (kv:its) -> Just (kv, Map.fromList its)
+
+-- * Internal
+positive :: Integral n => Series m n
+positive = generate $ \d -> [0..fromIntegral d]
+
+copositive :: (Num a, Ord a, CoSerial m a) => Series m b -> Series m (a -> b)
+copositive rs =
+    alts0 rs >>- \z ->
+    alts1 rs >>- \f ->
+    return $ \w ->
+      if w > 0
+        then f (w-1)
+        else z